Refactoring to LOAD DATA statements
